feat(code): improved code organization
